DEV Add utils.compare module
All comparison functions are included in
s/u/compare.py
The compareDocDecorator is included in s/u/docstrings.py

All are included in the __all__ lists at the top of the files
and included in the s/u/__init__.py
